# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个简单的机器管理系统",
        "title": "rabbit",
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/api/v1/host_group/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取所有host group信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.APIGetHostGroupListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_group/bind_host": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取host group信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "根据ID获取host group树状信息",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.HostGroupRel"
                        }
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_group/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "创建host group信息",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.APIGetHostGroupListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_group/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "删除host group信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "删除host group信息",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.HostGroup"
                        }
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_group/get": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "根据机器群组名称获取机器群组详细信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "根据机器群组名称获取机器群组详细信息",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_group/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取host group列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.APIGetHostGroupListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_group/related_hosts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "根据group路径获取group下所有的host",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.HostGroup"
                        }
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/host_group/update": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新host group信息",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "caasServiceId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "parentID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "parentName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/node.HostGroup"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "node.APIGetHostGroupListOutputs": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/node.HostGroup"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
        "node.HostGroup": {
            "type": "object",
            "properties": {
                "caasServiceId": {
                    "type": "integer"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/node.HostGroup"
                    }
                },
                "childrenHostCount": {
                    "type": "integer"
                },
                "childrenPodCount": {
                    "type": "integer"
                },
                "createUser": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isWarning": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parentID": {
                    "type": "integer"
                },
                "parentName": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "pathArray": {
                    "type": "string"
                },
                "relatedHostCount": {
                    "type": "integer"
                },
                "relatedPodCount": {
                    "type": "integer"
                },
                "subGroupCount": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "node.HostGroupRel": {
            "type": "object",
            "properties": {
                "groupID": {
                    "type": "integer"
                },
                "hostID": {
                    "type": "integer"
                }
            }
        },
...
    }
}